requirements.txt
setuptools
cffi
cryptography
azure-storage-queue==2.1.0
requests
